requirements.txt
Adafruit-BBIO==1.0.1
Adafruit-Libraries==0.1
pyserial==3.2.1
